Form bugfix

+++ src/org/moosbusch/lumPi/gui/form/editor/binding/spi/AbstractButtonFormEditorBindMapping.java
@@ -24,6 +24,7 @@ import org.moosbusch.lumPi.gui.form.editor.FormEditor;
 /**
  *
  * @author moosbusch
+ * @param <T>
  */
 public abstract class AbstractButtonFormEditorBindMapping<T extends Button>
         extends AbstractFormEditorBindMapping<T>
@@ -100,11 +101,11 @@ public abstract class AbstractButtonFormEditorBindMapping<T extends Button>
                 }
             } else if (value instanceof Number) {
                 Number n = (Number) value;
-                if (n == -1) {
+                if (n.intValue() == -1) {
                     return State.UNSELECTED;
-                } else if (n == 0) {
+                } else if (n.intValue() == 0) {
                     return State.MIXED;
-                } else if (n == 1) {
+                } else if (n.intValue() == 1) {
                     return State.SELECTED;
                 }
             } else if (value instanceof String) {
